requirements.txt
flask==2.3.3
werkzeug==2.3.7
google-generativeai>=0.3.1,<0.4.0
beautifulsoup4==4.9.3
requests>=2.28.0
python-dotenv==0.19.0
chromadb==0.4.22
langchain==0.1.0
langchain-google-genai==0.0.5 